requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
# created-by: conda 25.5.1
aiohappyeyeballs=2.6.1=pypi_0
aiohttp=3.12.15=pypi_0
aiosignal=1.4.0=pypi_0
async-timeout=5.0.1=pypi_0
attrs=25.3.0=pypi_0
bzip2=1.0.8=h2bbff1b_6
ca-certificates=2025.9.9=haa95532_0
certifi=2025.10.5=pypi_0
charset-normalizer=3.4.3=pypi_0
colorama=0.4.6=pypi_0
datasets=4.1.1=pypi_0
dill=0.4.0=pypi_0
expat=2.7.1=h8ddb27b_0
filelock=3.19.1=pypi_0
frozenlist=1.7.0=pypi_0
fsspec=2025.9.0=pypi_0
hf-xet=1.1.10=pypi_0
huggingface-hub=0.35.3=pypi_0
idna=3.10=pypi_0
libffi=3.4.4=hd77b12b_1
libzlib=1.3.1=h02ab6af_0
multidict=6.6.4=pypi_0
multiprocess=0.70.16=pypi_0
numpy=2.2.6=pypi_0
openssl=3.0.18=h543e019_0
packaging=25.0=pypi_0
pandas=2.3.3=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyhc872135_0
propcache=0.4.0=pypi_0
pyarrow=21.0.0=pypi_0
python=3.10.18=h981015d_0
python-dateutil=2.9.0.post0=pypi_0
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
requests=2.32.5=pypi_0
setuptools=78.1.1=py310haa95532_0
six=1.17.0=pypi_0
sqlite=3.50.2=hda9a48d_1
tk=8.6.15=hf199647_0
tqdm=4.67.1=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025.2=pypi_0
ucrt=10.0.22621.0=haa95532_0
urllib3=2.5.0=pypi_0
vc=14.3=h2df5915_10
vc14_runtime=14.44.35208=h4927774_10
vs2015_runtime=14.44.35208=ha6b5a95_10
wheel=0.45.1=py310haa95532_0
xxhash=3.6.0=pypi_0
xz=5.6.4=h4754444_1
yarl=1.20.1=pypi_0
zlib=1.3.1=h02ab6af_0
